[KED-1887] Refactoring on sidebar updates
