--- src/components/node-list/filter-nodes.js
@@ -1,84 +1,0 @@
-import { createSelector } from 'reselect';
-import utils from '@quantumblack/kedro-ui/lib/utils';
-const { escapeRegExp, getHighlightedText } = utils;
-
-/**
- * Get a list of IDs of the visible nodes
- * @param {object} nodes Grouped nodes
- * @return {array} List of node IDs
- */
-export const getNodeIDs = nodes => {
-  const getNodeIDs = type => nodes[type].map(node => node.id);
-  const concatNodeIDs = (nodeIDs, type) => nodeIDs.concat(getNodeIDs(type));
-
-  return Object.keys(nodes).reduce(concatNodeIDs, []);
-};
-
-/**
- * Add a new highlightedLabel field to each of the node objects
- * @param {object} nodes Grouped lists of nodes
- * @param {string} searchValue Search term
- * @return {object} The grouped nodes with highlightedLabel fields added
- */
-export const highlightMatch = (nodes, searchValue) => {
-  const addHighlightedLabel = node => ({
-    highlightedLabel: getHighlightedText(node.name, searchValue),
-    ...node
-  });
-  const addLabelsToNodes = (newNodes, type) => ({
-    ...newNodes,
-    [type]: nodes[type].map(addHighlightedLabel)
-  });
-
-  return Object.keys(nodes).reduce(addLabelsToNodes, {});
-};
-
-/**
- * Check whether a name matches the search text
- * @param {string} name
- * @param {string} searchValue
- * @return {boolean} True if match
- */
-export const nodeMatchesSearch = (node, searchValue) => {
-  const valueRegex = searchValue
-    ? new RegExp(escapeRegExp(searchValue), 'gi')
-    : '';
-  return Boolean(node.name.match(valueRegex));
-};
-
-/**
- * Return only the results that match the search text
- * @param {object} nodes Grouped lists of nodes
- * @param {string} searchValue Search term
- * @return {object} Grouped nodes
- */
-export const filterNodes = (nodes, searchValue) => {
-  const filterNodesByType = type =>
-    nodes[type].filter(node => nodeMatchesSearch(node, searchValue));
-  const filterNodeLists = (newNodes, type) => ({
-    ...newNodes,
-    [type]: filterNodesByType(type)
-  });
-
-  return Object.keys(nodes).reduce(filterNodeLists, {});
-};
-
-/**
- * Return filtered/highlighted nodes, and filtered node IDs
- * @param {object} nodes Grouped lists of nodes
- * @param {string} searchValue Search term
- * @return {object} Grouped nodes, and node IDs
- */
-export const getFilteredNodes = createSelector(
-  [state => state.nodes, state => state.searchValue],
-  (nodes, searchValue) => {
-    const filteredNodes = filterNodes(nodes, searchValue);
-
-    return {
-      filteredNodes: highlightMatch(filteredNodes, searchValue),
-      nodeIDs: getNodeIDs(filteredNodes)
-    };
-  }
-);
-
-export default getFilteredNodes;
